+++ src/main/java/com/takumi/takumimod/entity/goal/FlyAtTargetGoal.java
@@ -23,7 +23,7 @@ public class FlyAtTargetGoal extends Goal
     private static final double FLIGHT_SPEED = 0.5D;
     private static final int MAX_FLIGHT_TICKS = 60;
     private static final int LANDING_SLOW_FALLING_TICKS = 100;
-    private static final int ATTACK_INTERVAL_TICKS = 20;
+    private static final int ATTACK_INTERVAL_TICKS = 10;
 
     private final CardboardBoxEntity mob;
     private LivingEntity target;